+++ frontend/src/contexts/CreateNetContext.tsx
@@ -25,6 +25,7 @@ export interface CreateNetContextValue {
   description: string; setDescription: Dispatch<SetStateAction<string>>;
   infoUrl: string; setInfoUrl: Dispatch<SetStateAction<string>>;
   streamUrl: string; setStreamUrl: Dispatch<SetStateAction<string>>;
+  logoUrl: string | null; setLogoUrl: Dispatch<SetStateAction<string | null>>;
   script: string; setScript: Dispatch<SetStateAction<string>>;
   announcements: string; setAnnouncements: Dispatch<SetStateAction<string>>;